feat: 리뷰 단일 조회 API 구현

+++ src/main/java/com/dataury/soloJ/domain/touristSpot/entity/TouristSpot.java
@@ -22,7 +22,7 @@ public class TouristSpot extends BaseEntity {
     private String name;
 
     @Column(nullable = false)
-    private Long contentTypeId;
+    private int contentTypeId;
 
     @Column(nullable = false)
     private double latitude;
